Rework Seed Data Logic
- Rename correspondence seeds files with correspondences in name
in order to clarify their purpose.

- Create seed file for creating correspondence auto-assignment
levers.
