+++ db/seeds.rb
@@ -39,11 +39,13 @@ class SeedDB
 
     call_and_log_seed_step Seeds::Annotations
     call_and_log_seed_step Seeds::Tags
+
     # These must be ran before others
     call_and_log_seed_step Seeds::BusinessLineOrg
     call_and_log_seed_step Seeds::Users
     call_and_log_seed_step Seeds::NotificationEvents
     # End of required to exist dependencies
+
     call_and_log_seed_step Seeds::Tasks
     call_and_log_seed_step Seeds::Hearings
     call_and_log_seed_step Seeds::Intake
@@ -67,13 +69,15 @@ class SeedDB
     call_and_log_seed_step Seeds::MultiCorrespondences
     call_and_log_seed_step Seeds::VbmsDocumentTypes
     call_and_log_seed_step Seeds::VhaChangeHistory
+    call_and_log_seed_step Seeds::RemandedAmaAppeals
+    call_and_log_seed_step Seeds::RemandedLegacyAppeals
+    call_and_log_seed_step Seeds::CorrespondenceAutoTexts
+    call_and_log_seed_step Seeds::CorrespondenceAutoAssignmentLevers
+    call_and_log_seed_step Seeds::CorrespondenceAutoAssign
+
     # Always run this as last one
     call_and_log_seed_step Seeds::StaticTestCaseData
     call_and_log_seed_step Seeds::StaticDispatchedAppealsTestData
-    call_and_log_seed_step Seeds::AutoTexts
-    call_and_log_seed_step Seeds::RemandedAmaAppeals
-    call_and_log_seed_step Seeds::RemandedLegacyAppeals
-    call_and_log_seed_step Seeds::AutoAssign
   end
 end
